configuration.md: clarify public option
The previous text for the `public` option implied that setting the
value to `true` would create an instance of Wiki.js that would be
accessible without needing to create an account. However, it only
provides the _ability_ to enable public accessibility by creating a
Guest user whose access needs to be configured within the wiki's Users
Settings. By default, the Guest user is denied read access. The help
text for the `public` option has been clarified to reflect this.